requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
python-dotenv==0.19.2
google-generativeai
